requirements.txt
google-cloud-bigquery==0.29.0
functions-framework==3.*
requests==2.31